Move radarview as an independant plot tool
